# openapi.yaml
openapi: 3.0.3
info:
  title: AndX Exchange Service API
  description: |
    AndX Exchange Service, kripto para piyasa verilerini sağlayan RESTful API servisidir. 
    CoinGecko ve Alternative.me API'lerini kullanarak gerçek zamanlı market verilerini sunar.

    ## Özellikler
    - Gerçek zamanlı kripto para market verileri
    - 90+ kripto para desteği
    - TRY ve USD para birimi desteği
    - Korku & Hırs endeksi
    - Rate limiting ve caching
    - Comprehensive error handling

  version: 1.0.0
  contact:
    name: AndX Tech Team
    email: support@andxtech.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server

paths:
  /market/{symbol}:
    get:
      tags:
        - Market Data
      summary: Kripto para market verilerini getirir
      description: |
        Belirtilen kripto para için detaylı market bilgilerini getirir.
        Veriler CoinGecko ve Alternative.me API'lerinden alınır.
      parameters:
        - name: symbol
          in: path
          required: true
          description: Kripto para sembolü (btc, eth, usdt, vb.)
          schema:
            type: string
            example: usdt
        - name: fiat
          in: query
          required: false
          description: Fiat para birimi
          schema:
            type: string
            enum: [try, usd]
            default: try
            example: try
      responses:
        '200':
          description: Başarılı response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
              examples:
                usdt_try:
                  summary: USDT (TRY)
                  value:
                    symbol: 'USDT'
                    coin_id: 'tether'
                    rank: 4
                    market_cap_try: 6834785722634
                    market_cap_usd: 167017182969
                    volume_24h_usd: 100722116342
                    circulating_supply: 167010130246.28
                    change_pct_7d: -0.004
                    change_pct_30d: -0.035
                    change_pct_90d: -0.016
                    ath_change_pct_usd: -24.42
                    atl_change_pct_usd: 74.67
                    fear_greed: 44
                    links:
                      whitepaper: 'https://tether.to/wp-content/uploads/2016/06/TetherWhitePaper.pdf'
                      website: 'https://tether.to/'
                      twitter: 'https://twitter.com/Tether_to'
                btc_usd:
                  summary: Bitcoin (USD)
                  value:
                    symbol: 'BTC'
                    coin_id: 'bitcoin'
                    rank: 1
                    market_cap_try: 92562894018804
                    market_cap_usd: 2262703217539
                    volume_24h_usd: 45225019762
                    circulating_supply: 19909075
                    change_pct_7d: -6.69
                    change_pct_30d: -3.54
                    change_pct_90d: 2.14
                    ath_change_pct_usd: -8.56
                    atl_change_pct_usd: 167282.97
                    fear_greed: 44
                    links:
                      whitepaper: 'https://bitcoin.org/bitcoin.pdf'
                      website: 'http://www.bitcoin.org'
                      twitter: 'https://twitter.com/bitcoin'
        '400':
          description: Geçersiz istek parametresi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 400
                timestamp: '2025-08-20T18:50:00.000Z'
                path: '/market/btc'
                method: 'GET'
                error: 'bad_request'
                message: 'Invalid fiat currency'
                requestId: '12345678-1234-5678-9abc-123456789012'
        '404':
          description: Kripto para bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                timestamp: '2025-08-20T18:51:06.466Z'
                path: '/market/invalidcoin'
                method: 'GET'
                error: 'not_found'
                message: "Cryptocurrency 'invalidcoin' not found"
                requestId: '47196144-0250-4b55-a9a8-5dc38df53797'
        '429':
          description: Rate limit aşıldı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 429
                timestamp: '2025-08-20T18:52:15.123Z'
                path: '/market/btc'
                method: 'GET'
                error: 'too_many_requests'
                message: 'Rate limit exceeded. Try again later.'
                requestId: '12345678-1234-5678-9abc-123456789012'
        '502':
          description: Upstream API hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 502
                timestamp: '2025-08-20T18:53:20.789Z'
                path: '/market/eth'
                method: 'GET'
                error: 'upstream_error'
                message: 'CoinGecko API error: Request timeout'
                requestId: '87654321-4321-8765-cba9-876543210987'

  /market/bulk:
    post:
      tags:
        - Market Data
      summary: Çoklu kripto para market verilerini getirir
      description: |
        Belirtilen kripto para listesi için detaylı market bilgilerini getirir.
        Veriler CoinGecko ve Alternative.me API'lerinden alınır.

        **Özellikler:**
        - En az 1 coin sembolü gereklidir
        - Başarısız olan coinler 'failed' listesinde döner
        - Başarılı olan coinler 'data' listesinde döner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMarketRequest'
            examples:
              basic:
                summary: Temel kullanım
                value:
                  symbols: ['btc', 'eth', 'usdt']
                  fiat: 'try'
              usd:
                summary: USD ile
                value:
                  symbols: ['btc', 'eth']
                  fiat: 'usd'
      responses:
        '200':
          description: Başarılı response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkMarketResponse'
        '400':
          description: Geçersiz parametre - Liste boş olamaz
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 400
                  error:
                    type: string
                    example: 'validation_error'
                  message:
                    type: string
                    example: 'Liste boş olamaz, en azından 1 coin olmalı'
        '429':
          description: Rate limit aşıldı
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 429
                  message:
                    type: string
                    example: 'Rate limit exceeded. Try again later.'

  /health:
    get:
      tags:
        - Health Check
      summary: API sağlık durumu
      description: API'nin çalışma durumunu ve uptime bilgisini döner
      responses:
        '200':
          description: API sağlıklı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: 'ok'
                timestamp: '2025-08-20T18:49:31.456Z'
                uptime: 23.09

components:
  schemas:
    MarketResponse:
      type: object
      properties:
        symbol:
          type: string
          description: Kripto para sembolü (büyük harf)
          example: 'USDT'
        coin_id:
          type: string
          description: CoinGecko ID
          example: 'tether'
        rank:
          type: integer
          nullable: true
          description: Market cap sıralaması
          example: 4
        market_cap_try:
          type: number
          nullable: true
          description: Piyasa değeri (TRY)
          example: 6834785722634
        market_cap_usd:
          type: number
          nullable: true
          description: Piyasa değeri (USD)
          example: 167017182969
        volume_24h_usd:
          type: number
          nullable: true
          description: 24 saatlik hacim (USD)
          example: 100722116342
        circulating_supply:
          type: number
          nullable: true
          description: Dolaşımdaki arz
          example: 167010130246.28
        total_supply:
          type: number
          nullable: true
          description: Toplam arz
          example: 167010130246.28
        max_supply:
          type: number
          nullable: true
          description: Maksimum arz
          example: 21000000
        change_pct_7d:
          type: number
          nullable: true
          description: 7 günlük değişim (%)
          example: -0.004
        change_pct_30d:
          type: number
          nullable: true
          description: 30 günlük değişim (%)
          example: -0.035
        change_pct_90d:
          type: number
          nullable: true
          description: 90 günlük değişim (%)
          example: -0.016
        ath_change_pct_usd:
          type: number
          nullable: true
          description: ATH'den değişim (%)
          example: -24.42
        atl_change_pct_usd:
          type: number
          nullable: true
          description: ATL'den değişim (%)
          example: 74.67
        fear_greed:
          type: integer
          nullable: true
          description: Korku & Hırs endeksi (0-100)
          example: 44
        links:
          $ref: '#/components/schemas/Links'
      required:
        - symbol
        - coin_id
        - links

    Links:
      type: object
      properties:
        whitepaper:
          type: string
          nullable: true
          description: Whitepaper URL
          example: 'https://tether.to/wp-content/uploads/2016/06/TetherWhitePaper.pdf'
        website:
          type: string
          nullable: true
          description: Resmi website URL
          example: 'https://tether.to/'
        twitter:
          type: string
          nullable: true
          description: Twitter profil URL
          example: 'https://twitter.com/Tether_to'

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: API durumu
          example: 'ok'
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2025-08-20T18:49:31.456Z'
        uptime:
          type: number
          description: Uptime (saniye)
          example: 23.09
      required:
        - status
        - timestamp
        - uptime

    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status kodu
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Hata timestamp'i
          example: '2025-08-20T18:51:06.466Z'
        path:
          type: string
          description: İstek path'i
          example: '/market/invalidcoin'
        method:
          type: string
          description: HTTP metodu
          example: 'GET'
        error:
          type: string
          description: Hata tipi
          example: 'not_found'
        message:
          type: string
          description: Hata mesajı
          example: "Cryptocurrency 'invalidcoin' not found"
        requestId:
          type: string
          description: Benzersiz istek ID'si
          example: '47196144-0250-4b55-a9a8-5dc38df53797'
      required:
        - statusCode
        - timestamp
        - path
        - method
        - error
        - message
        - requestId

    BulkMarketRequest:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
          minItems: 1
          description: Kripto para sembolleri listesi
          example: ['btc', 'eth', 'usdt']
        fiat:
          type: string
          enum: [try, usd]
          default: try
          description: Fiat para birimi
          example: 'try'
      required:
        - symbols

    BulkMarketResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MarketResponse'
          description: Market verileri listesi
        count:
          type: integer
          description: Toplam coin sayısı
          example: 3
        failed:
          type: array
          items:
            type: string
          description: Başarısız olan coinler
          example: ['invalidcoin']
      required:
        - data
        - count
        - failed

  securitySchemes: {}

tags:
  - name: Market Data
    description: Kripto para market verileri
  - name: Health Check
    description: API sağlık kontrolü